Refactor app code and Terraform deployment

+++ aws/serverless/_deploy.bash
@@ -7,22 +7,29 @@ if [ -z $action ]; then
     exit 1
 fi
 
-appDir="../app"
-apiDir="$appDir/api"
-staticDir="$appDir/static"
+staticDir="../../app/static"
+lambdaDir="./lambda"
 lambdaExecutable="handler"
+lambdaHashFile=".lambda_hash"
 
-export TF_VAR_lambda_zip="$apiDir/$lambdaExecutable.zip"
+# Set Terraform environment variables
+export TF_VAR_lambda_zip="$lambdaDir/$lambdaExecutable.zip"
 export TF_VAR_lambda_handler=$lambdaExecutable
 export TF_VAR_public_ip=$(curl -s https://checkip.amazonaws.com)
 export TF_VAR_static_files_dir=$staticDir
 
 if [ $action == "create" ]; then
-    # Build Go Lambda function handler...
-    cd $apiDir
-    GOOS=linux GOARCH=amd64 go build -o $lambdaExecutable "$lambdaExecutable.go"
-    zip "$lambdaExecutable.zip" $lambdaExecutable
-    rm -f $lambdaExecutable
+    # Calculate whether the Go Lambda function handler needs to be built again
+    cd $lambdaDir
+    lastHash=$(cat $lambdaHashFile || echo "")
+    currentHash=$(cat "$lambdaExecutable.go" | openssl dgst -sha256)
+
+    if [ "$lastHash" != "$currentHash" ]; then
+        GOOS=linux GOARCH=amd64 CGO_ENABLED=0 go build -o $lambdaExecutable "$lambdaExecutable.go"
+        zip "$lambdaExecutable.zip" $lambdaExecutable
+        rm -f $lambdaExecutable
+        echo -n $currentHash > $lambdaHashFile
+    fi
 
     # Deploy to AWS with Terraform
     cd -
